compute*: introduce an UpdateConfiguration command
This commit replaces the `UpdateMaxResultSize` command with a more
general `UpdateConfiguration` command that will also allow updating
other configuration options in the future.

The `UpdateConfiguration` command represents config updates as a
`BTreeSet<ComputeParameter>` rather than a single `ComputeConfig` struct
with `Option` fields, to ensure that updates of single parameters have a
space-efficient representation even as the set of compute parameters
grows large.