cloudtest,mzcompose: generalize external service configuration
This commit avoids duplicating the configuration of CockroachDB and
MinIO across our various test harnesses. The mzcompose `Cockroach`,
`MinioSetup`, and `Materialized` services learn convenience options that
configure them all appropriately for one another. Additionally, the
CockroachDB initialization is shared between cloudtest and mzcompose.
This yields a natural spot to install the workaround for the CockroachDB
bug discovered in #16726.

Fix #16726.