sql: temporarily allow "mzcompose" pseudo cloud provider
fbb31a61 attempted to replace the "mzcompose" pseudo cloud provider with
a "docker" psuedo cloud provider. The motivation is to have a single
cloud provider for use when running the Docker image directly, via
`docker run` and when running via mzcompose.

Unfortunately, changing the environment ID conditionally based on the
version of the Materialize in use breaks the upgrade tests (see #16902).
So instead we'll need to do this over two releases. In the next release,
we'll add support for the "docker" provider without removing support for
the "mzcompose" provider; in the *following* release,  we'll remove
support for the "mzcompose" provider.