Expecting elements takes closure
Previously, expect_element would accept a fully-formed message, which meant
the message would need to be fully computed when calling the function. For
specialized messages, this would cause an allocation and string
interpolation.

With this change, expect_element takes a closure that is used in the error
case to generate a message. This avoids pre-generating error messages at
the expense of some additional syntactic noise.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>