mzcompose: Protect --persist-blob-url from the shell
previously, the --persist-blob-url option was passed as an environment
variable. Now it is passed on the command line, which means it
needs to be single-quoted in case it contains special shell characters.

For example, if the --persist-blob-url was pointing to a Minio instance,
the URL will contain the string '&region=minio' . The ampresand in
this string would cause the rest of the command line to be disregarded.