environmentd: add a readiness probe
Add an readiness probe on the internal HTTP server at /api/readyz that
reports whether Materialize is ready to respond to requests. It reports
a 503 status while the server is booting, and a 200 status thereafter.
The `?wait=true` query parameter can be used to request that the
response block until the server becomes ready.

This probe will be immediately useful as a Kubernetes readiness probe
and a Docker Compose healthcheck.