pgrepr: adopt UInt wrapper types
Move the Uint2, Uint4, and Uint8 wrapper types from testdrive to pgrepr,
where their implementation can be shared across the various crates that
were reinventing the wheel (sqllogictest, pgrepr, testdrive, and
mzcloud_shared in the cloud repository).

Also change their spelling to "UInt" to matcdh the spelling in the rest
of the codebase.