test/upgrade: use standalone CockroachDB
The upgrade test would inadvertently test the upgrade/downgrade of
CockroachDB in addition to the upgrade/downgrade of Materialize if we
upgraded the version of CockroachDB embedded in the materialized
container.  If we ever *downgraded* the embedded CockroachDB, the
upgrade tests would fail, since upgrading Materialize would result in
downgrading Cockroach, which is unsupported.

This commit changes the test to use a standalone CockroachDB that does
not get upgraded/downgraded alongside Materialize.