compute history: fix UpdateConfiguration reduction
Prior to this commit, history reduction for `UpdateConfiguration`
commands was broken. We simply collected config parameters in a
`BTreeSet`, which would *not* deduplicate parameters of the same type.
As a result, new replicas would receive the sum of all configuration
updates ever made and would apply them all.

This is fixed by using a `BTreeMap` instead, keyed by the config
parameter key.