Initial version of COMPUTE Layer Q&A (#17048)
This commit adds an initial version of a Q&A document on the COMPUTE layer to our developer documentation regarding Materialize Platform. The Q&A is still incomplete in the sense that it has more questions than answers filled in. In contrast to an initial high energy editing pass to get the questions in place and a few initial answers, the document is now in a stage of its lifecycle that is more adequate for review and contributions via PRs than via internal editing tools.