persist: retry Blob::open and Consensus::open in maelstrom
This was causing failures in nightly. We do these retries in prod
codepaths, but because of Unreliable{Blob,Consensus} and
Maelstrom{Blob,Consensus} this wasn't going through the same codepath. A
better fix would be to figure out how to unify things here, but as a
short-term fix to unbreak nightly, this is the quicker thing.