slt: make persisted introspection configurable
Prior to this change, sqllogictest always enabled persisted
introspection sources. We observed that this had a significant impact on
the time it takes to run tests, likely caused by the increased persist
usage.

This commite adds a flag, `--persisted-introspection`, to the
sqllogictest binary that enables configuring whether persisted
introspection sources should be switched on or not. By default,
persisted introspection is disabled.

This commit also moves the slt tests that test introspection sources
into a separate subfolder. This allows us to configure the CI scripts to
only enable persisted introspection when testing introspection sources,
and disable it otherwise.