test/restart: fortify stash test
The stash test could not handle all the ways in which the Materialize
container could crash when the CockroachDB database to which it is
talking is wiped.

This commit restructures the test to be a bit more resilient. I ran this
test in a loop on my machine for an hour with no observed failures.

The key here is keeping a connection to Materialize open during the
whole test. We *know* that connection will succeed initially
(Materialize is healthy); we *know* it will fail once we wipe Cockroach.
Keeping the connection open the whole time means we don't have to deal
with myriad failure conditions (connection refused, `docker port`
failing), and can instead just look for pg8000 reporting a network error
on the broken connection.

This commit also removes the "stash error" as an allowable error; in the
one hour I ran this test I never saw that error. Thinking about it more,
I'm not sure how it could happen. While Cockroach is down, the stash
should retry forever; when Cockroach comes back up and the stash
reconnections, it should immediately panic when it discovers that the
tables it previously installed are missing.