Propagate debug_name instead of dataflow_id in reduce rendering (#16998)
This commit propagates the more informative dataflow debug name rather than its id to aid in debugging errors originating from reduce processing of invalid accumulations. This is not the final state we would like the code to be left at as, but rather an interim measure until more extensive refactoring can be performed.