compute-client: error targeted subscribes on replica drop
When the compute controller observes that a replica that is targeted by
subscribes disconnects or is dropped by the user, it cancels those
subscribes by sending an error response upstream. The error response has
the effect that the coordinator cancels the subscribe and displays the
error to the user.

We need to cancel replica-targeted subscribes even if the target replica
only temporarily disconnects and is re-hydrated successfully. This is
because we don't know after a reconnect if the replica has restarted or
not. If it has restarted, it has created new introspection collections
that have different contents than the ones subscribes may have observed
previously. It would be unsound to produce the contents of two different
collections in the same subscribe, so the subscribe must be shut down.