storage: reduce the size of DataflowError
This commit boxes all the variants of `DataflowError`, to reduce the
type's size in memory. This reduces the size of `Result<Row,
DataflowError>` values from 80 bytes (the previous size of
`DataflowError`) to 32 bytes (the size of `Row`).

The benefit of this change is that it removes the risk of us
accidentially wasting a lot of memory by using above `Result` type on
high-throughput code paths.