storage: introduce independent storage instance management
Part of the cluster unification epic (MaterializeInc/cloud#4929).

This commit introduces the concept of a independently-managable storage
instance to the storage controller. Previously, a storage instance was
always 1:1 with an ingestion or an export; now, a storage instance is
created independently of any ingestion or export, and can in principle
(though not yet in practice) contain any number of ingestions or
exports.

Co-authored-by: Gus Wynn <gus@materialize.com>