adapter: consistentize Storage::remove_cluster API
Part of the cluster unification epic (MaterializeInc/cloud#4929).

For consistency with the other catalog storage APIs, this function
should take the cluster ID as input and return nothing as output.

Something is still fishy here--the storage layer should not be
responsible for DROP ... CASCADE behavior--but leaving that
investigation to another day.