Kill all of sqllogictest's descendants
Currently, sqllogictest will leave clusterd processes around when it
either crashes or fails -- in the latter case, this is because it
process::exit()s on failure, which causes the Drop
impl (KillOnDropChild) in the process orchestrator to not fire.

The reason sqllogictest doesn't reparent these like environmentd is
that it generates a new environment-id every run; I prepared an
alternate branch where I do that instead, but it seems more
self-contained to just kill everything in the process group.

Note that it's not sufficient to kill descendants via PPID, because
the clusterd processes have already been reparented to init by the
time we know we need to clean up.