optimizer: forbid use of stdlib Hash collections
This commit enables clippy lints to deny the use of the stdlib `Hash`
collections in crates that are already not using them anymore.

The final goal is to enable the `clippy::disallowed_types` lint
globally, once all crates have been cleansed of `Hash` collection uses.
At that point we will be able to remove the manual lint enables added in
this commit again.