storage: only create async worker once
Similar to the internal command sequencer, it is very important that we
only create the async worker once because a) the `StorageState` is
re-used when a new client connects and b) commands that have already
been sent and might yield a response will be lost if a new iteration of
`run_client` creates a new async worker.

If we created a new async worker every time we get a new client (likely
because the controller re-started and re-connected), we can get into an
inconsistent state where we think that a dataflow has been rendered, for
example because there is an entry in `StorageState::ingestions`, while
there is not yet a dataflow. This happens because the dataflow only gets
rendered once we get a response from the async worker and send off an
internal command.

The core idea is that both the sequencer and the async worker are part
of the per-worker state, and must be treated as such, meaning they must
survive between invocations of `run_client`.

Fixes #17309