Bump Rust version to 1.67
This requires us to silence a new clippy lint,
`clippy::unchecked_duration_subtraction`. While the lint triggering is
not an FP, it is not very helpful either in this case. Substracting a
small tick duration from the current time will never underflow, and if
it did we probably would like to panic anyway. Note that the lint seems
to be "allow by default" on the current clippy main, so we should be
able to remove the explicit #[allow] again in the future.