cloud-canary: Use the same image for both local and cloud Mz instances
Use the same Materialized image DockerHub tag for the local and
the cloud Mz instances. Otherwise mzcompose may pick a different
commit for the local Mz instance and the outputs of mz_version()
will no longer be comparable.

This happens if there has been a commit, e.g. documentation, that
did not touch the product code per-se.