adapter: Fix bug in initialize_read_policies (#17421)
Previously, when initializing read policies, we were grouping all
IdBundles by the timestamp of their read holds, in a HashMap. Instead
of extending the keys of the HashMaps, we were overwriting them.
So we would end up forgetting all but the last occurrence of each
timestamp. This commit fixes this issue by extending each key of the
HasMap, instead of overwriting it.