storage: make sure we efficiently batch records from sources
Due to a common pattern in async operators, namely activating the output
handle, creating a session, sending a single message, and immediately
deactivating the code in `main` was creating one batch per message which
had the effect of allocating a ton of small vectors and making
performance much worse.

This pattern is used because there is an impedence mismatch between how
output handles work and how async operators work. Ideally what we want
is a handle that will buffer messages and automatically flushes when the
async future gets into a `Pending` state.

This handle doesn't exist (yet) and it's complicated to write.

This PR fixes the immediate issue by rearranging the code so that output
handle activation can be done once for all the messages that are
currently ready and leaves the specialized async output handle for a
Future(ha) date.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>